+++ tests/PostsManager.addPost.test.tsx
@@ -1,0 +1,107 @@
+import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest"
+import { render, screen, waitFor } from "@testing-library/react"
+import userEvent from "@testing-library/user-event"
+import { http, HttpResponse } from "msw"
+import { setupServer } from "msw/node"
+import { MemoryRouter } from "react-router-dom"
+import PostsManager from "../src/pages/PostsManagerPage"
+import * as React from "react"
+import "@testing-library/jest-dom"
+import { TEST_POSTS, TEST_SEARCH_POST, TEST_USERS } from "./mockData"
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+
+// MSW 서버 설정
+const server = setupServer(
+  http.get("/api/posts", () => {
+    return HttpResponse.json(TEST_POSTS)
+  }),
+
+  http.get("/api/posts/search?q=His%20mother%20had%20always%20taught%20him", () => {
+    return HttpResponse.json(TEST_SEARCH_POST)
+  }),
+
+  http.get("/api/users", () => {
+    return HttpResponse.json(TEST_USERS)
+  }),
+
+  http.get("/api/posts/tags", () => {
+    return HttpResponse.json([
+      "history",
+      "american",
+      "crime",
+      "french",
+      "fiction",
+      "english",
+      "magical",
+      "mystery",
+      "love",
+      "classic",
+    ])
+  }),
+)
+
+beforeAll(() => server.listen({ onUnhandledRequest: "error" }))
+afterEach(() => server.resetHandlers())
+afterAll(() => server.close())
+
+const renderPostsManager = () => {
+  const queryClient = new QueryClient()
+  return render(
+    <QueryClientProvider client={queryClient}>
+      <MemoryRouter>
+        <PostsManager />
+      </MemoryRouter>
+    </QueryClientProvider>,
+  )
+}
+
+describe("PostsManager", () => {
+  it("새 게시물 추가를 허용합니다", async () => {
+    const user = userEvent.setup()
+    const NEW_POST = {
+      id: TEST_POSTS.posts.length + 1,
+      title: "New Post",
+      body: "This is a new post",
+      userId: 1,
+      tags: [],
+    }
+
+    // POST 요청에 대한 핸들러 추가
+    server.use(
+      http.post("/api/posts/add", async ({ request }) => {
+        const body = await request.json()
+        // 요청 body 검증 (선택적)
+        expect(body).toMatchObject({
+          title: NEW_POST.title,
+          body: NEW_POST.body,
+        })
+        return HttpResponse.json(NEW_POST)
+      }),
+    )
+
+    renderPostsManager()
+
+    // 기존 게시물들이 로드될 때까지 대기
+    await waitFor(() => {
+      TEST_POSTS.posts.forEach((post) => {
+        expect(screen.getByText(post.title)).toBeInTheDocument()
+      })
+    })
+
+    const addButton = screen.getByRole("button", { name: /게시물 추가/i })
+    await user.click(addButton)
+
+    const titleInput = screen.getByPlaceholderText(/제목/i)
+    const bodyInput = screen.getByPlaceholderText(/내용/i)
+    await user.type(titleInput, NEW_POST.title)
+    await user.type(bodyInput, NEW_POST.body)
+
+    const submitButton = screen.getByRole("button", { name: /게시물 추가/i })
+    await user.click(submitButton)
+
+    // 새 게시물이 추가되었는지 확인
+    await waitFor(() => {
+      expect(screen.getByText(NEW_POST.title)).toBeInTheDocument()
+    })
+  })
+})
